+++ components/stories/CookieBanner.stories.tsx
@@ -1,9 +1,9 @@
 import React from 'react'
 
-import { CookieBanner } from '../CookieBanner'
+import { CookieBanner, SavedSettings } from '../CookieBanner'
 
 export const Default = () => {
-  return <CookieBanner />
+  return <CookieBanner value={{} as SavedSettings} setValue={() => null} />
 }
 
 // eslint-disable-next-line import/no-default-export
